Fix test to work for updated data
One of the unit tests generated by make_tests.py creates a validation failure
on purpose by altering some rdata, however the change was done by
search/replace, but the actual rdata has changed since the test was written, so
it breaks unit test generation. This patch makes the change in a different way
that will koop working when the RRset changes again.