Make deployment work on latest Debian (bookwork)
The README says the deployment should work on the latest Debian. They worked
fine on Debian 11 (bullseye), but a few changes needed to be made to make it
work on Debian 12 (bookworm). This includes removal of some package
installations (that weren't necessary anyway), and allowing pip3 to "break
system packages". Also more privileges were needed to deploy the collector.